Move components to separate files and tweak a few things
