+++ code/core/src/components/components/Form/Form.tsx
@@ -3,8 +3,10 @@ import { styled } from 'storybook/theming';
 import { Button } from '../Button/Button';
 import { Checkbox } from './Checkbox';
 import { Field } from './Field';
-import { Input, Select, Textarea } from './Input';
+import { Input } from './Input';
 import { Radio } from './Radio';
+import { Select } from './Select';
+import { Textarea } from './Textarea';
 
 export const Form = Object.assign(
   styled.form({
